# openapi.yaml
openapi: "3.0.3"
info:
  title: Engram Memory API
  description: |
    HTTP API for the Engram memory graph system.

    Engram stores memories as a four-tier graph:
    - **Episodes** (Tier 1): Raw ingested messages and events
    - **Entities** (Tier 2): Named entities extracted from episodes
    - **Engrams** (Tier 3): Consolidated memories built from episode clusters
    - **Schemas** (Tier 4): Cross-cutting pattern templates extracted from L2+ engrams

    All `/v1/*` endpoints require Bearer token authentication.
    The `/health` endpoint is public.

    ## Identity Configuration

    Set `identity` in `engram.yaml` to enable role-aware memory consolidation.
    When configured, the consolidation pipeline labels episode fragments by role
    (bot, owner, or third party) so memories are generated with correct attribution.

    ```yaml
    identity:
      name: "Bud"          # Bot display name; matches episode.author
      author_id: "bud"     # Matches episode.author_id set by callers
      owner_ids:            # Optional; enables owner-specific framing
        - "thunder"
    ```

    `POST /v1/thoughts` automatically stamps `author`/`author_id` from the
    identity config so the bot's internal monologue is correctly attributed.

    Env var overrides: `ENGRAM_IDENTITY_NAME`, `ENGRAM_IDENTITY_AUTHOR_ID`.
  version: "0.1.0"
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key passed as Bearer token. Only required when `server.api_key` is set
        in the server config. If no key is configured, auth is disabled and all
        `/v1/*` requests are accepted without a header.
  schemas:
...
    Schema:
      type: object
      description: |
        A cross-cutting pattern template extracted from L2+ engrams.
        Not a summary of what happened — a template for what class of event this is,
        with generalizations extracted from multiple instances.
      properties:
        id:
          type: string
          description: 32-char BLAKE3 hex ID; use any 5+ char prefix to reference
        name:
          type: string
          description: Short human-readable label (e.g. "Memory System Debugging")
        content:
          type: string
          description: |
            Full semi-structured prose with PATTERN, GENERALIZATIONS, INSTANCES, etc.
            When `?level=N` is requested, replaced by a precomputed N-word summary of
            the GENERALIZATIONS section.
        is_labile:
          type: boolean
          description: If true, needs reconsolidation at the next induction run
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        instances:
          type: array
          items:
            $ref: "#/components/schemas/SchemaInstance"
          description: Engrams that matched this schema (populated on retrieval)
    SchemaInstance:
      type: object
      description: Records that a specific engram matched a schema, with extracted slot values.
      properties:
        schema_id:
          type: string
        engram_id:
          type: string
        slot_values:
          type: object
          additionalProperties:
            type: string
          description: "Extracted slot values keyed by slot name (e.g. {\"trigger\": \"...\", \"fix\": \"...\"})"
        is_anomaly:
          type: boolean
          description: True if the instance deviates from the schema pattern
        matched_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          example: "missing_field"
paths:
...
  /v1/schemas:
    get:
      summary: List schemas
      description: |
        Returns all schemas ordered by `updated_at` descending.
        Embeddings are never returned. Content is the full semi-structured prose.
      operationId: listSchemas
      tags:
        - Schemas
      responses:
        "200":
          description: List of schemas
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    content:
                      type: string
                    is_labile:
                      type: boolean
                    created_at:
                      type: string
                      format: date-time
                    updated_at:
                      type: string
                      format: date-time
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/schemas/{id}:
    get:
      summary: Get a schema
      description: |
        Returns a single schema with its instance list.

        Add `?level=N` to replace the `content` field with a precomputed N-word summary
        of the GENERALIZATIONS section (the most stable part of the schema). Valid levels:
        4, 8, 16, 32, 64. If no precomputed summary exists at that level, `content` is
        returned verbatim.
      operationId: getSchema
      tags:
        - Schemas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Full schema ID or 5-char short ID
        - name: level
          in: query
          schema:
            type: integer
          description: "Return a precomputed summary at this compression level instead of full content"
      responses:
        "200":
          description: Schema with instance list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schema"
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Schema not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete a schema
      description: Permanently deletes a schema and its instance annotations.
      operationId: deleteSchema
      tags:
        - Schemas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Full schema ID or 5-char short ID
      responses:
        "200":
          description: Schema deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "deleted"
        "400":
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/schemas/induce:
    post:
      summary: Trigger schema induction
      description: |
        Runs the schema induction pipeline asynchronously over all L2+ engrams. Requires
        at least 3 L2+ engrams; returns `{"started": false}` otherwise.

        Induction clusters engrams by activity type/pattern, then uses an LLM to extract
        cross-cutting schemas. Requires `llm` to be configured in `engram.yaml`.

        Returns HTTP 202 immediately when started; induction continues in the background.
        Run this manually to backfill `schema_summaries` after first deployment, or after
        a DB wipe. Otherwise induction runs automatically as part of consolidation.
      operationId: induceSchemas
      tags:
        - Schemas
      responses:
        "200":
          description: Not enough L2+ engrams to run induction
          content:
            application/json:
              schema:
                type: object
                properties:
                  started:
                    type: boolean
                    example: false
                  reason:
                    type: string
                    example: "not enough L2+ engrams (need at least 3)"
        "202":
          description: Induction started asynchronously
          content:
            application/json:
              schema:
                type: object
                properties:
                  started:
                    type: boolean
                    example: true
        "500":
          description: Database error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Schema induction not configured (no LLM)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...